+++ electron-app/src/components/options/ToolsOptionsPanel.tsx
@@ -296,6 +296,7 @@ export function ToolsOptionsPanel({
               <code className="text-void-light">search_files</code>, <code className="text-void-light">glob_files</code>,{' '}
               <code className="text-void-light">git_status</code>, <code className="text-void-light">git_diff</code>,{' '}
               <code className="text-void-light">git_log</code>, <code className="text-void-light">git_show</code>,{' '}
+              <code className="text-void-light">check_types</code>,{' '}
               <code className="text-void-light">execute_command</code>.
             </>
           }